Done basic refactoring of codebase
