+++ src/index.ts
@@ -1,5 +1,4 @@
 import dotenv from "dotenv";
-import twilio, { Twilio } from "twilio";
 import express, { Request, Response } from "express";
 import bodyParser from "body-parser";
 import cors from "cors";
@@ -14,21 +13,13 @@ import voiceLabRoutes from "./routes/voiceLabRoutes";
 import fileRoutes from "./routes/fileRouter";
 import { loadSwaggerDocs } from "./swaggerLoader";
 import MessagingResponse from "twilio/lib/twiml/MessagingResponse";
-import OpenAI from "openai";
-const openai = new OpenAI();
-import WebSocketClient from "ws"; // Rename to WebSocketClient to avoid confusion
-import { PrismaClient } from "@prisma/client"; // Import Prisma Client
 
-import { createClient, LiveTranscriptionEvents } from "@deepgram/sdk";
 
-const deepgramClient = createClient(process.env.DEEPGRAM_API_KEY);
+import MediaStream from './services/MediaStream'
 
 // Initialize dotenv for environment variables
 dotenv.config();
 
-// Twilio credentials and setup
-const accountSid: string = process.env.TWILIO_ACCOUNT_SID || "";
-const authToken: string = process.env.TWILIO_AUTH_TOKEN || "";
 
 // Express app setup
 const app = express();
@@ -70,352 +61,6 @@ wss.on("connection", (ws: WebSocket) => {
   const clientConnection = new MediaStream(ws);
 });
 
-
-
-
-
-const prisma = new PrismaClient(); // Initialize Prisma Client
-
-let streamSid: string = "";
-
-
-class MediaStream {
-  private connection: WebSocket;
-  deepgram: any;
-  deepgramTTSWebsocket: WebSocket;
-  messages: never[] = [];
-  repeatCount: number = 0;
-  private AgentObject: any[] = []; // Array to store agent objects
-
-  constructor(ws: WebSocket) {
-    this.connection = ws;
-
-    // Ensure connection is properly initialized and ready to send messages
-    this.connection.on("open", () => {
-      console.log("WebSocket connection opened.");
-
-    });
-
-    this.connection.on("message", this.processData.bind(this));
-
-    this.deepgram = setupDeepgram(this);
-    this.deepgramTTSWebsocket = setupDeepgramWebsocket(this);
-  }
-
-
-
-  
-
-  processData = async (data: string) => {
- 
-
-
-    const DataObject = JSON.parse(data);
-    const event = DataObject.event;
-
-    if (event === "start") {
-      log("Call accepted");
-      streamSid = DataObject.streamSid;
-
-      const client = twilio(accountSid, authToken);
-      client
-        .calls(DataObject.start.callSid)
-        .fetch()
-        .then(async (call) => {
-          console.log(`To: ${call.to}`);
-
-          const agent = await prisma.agent.findFirst({
-            where: { phoneNumber: call.to },
-          });
-
-          if (agent) {
-            this.AgentObject.push(agent); // Push the agent object to the array
-          }
-
-
-          console.log(this.getAgentObject()[0].welcomeMessage)
-          promptLLM(this,this.getAgentObject()[0].welcomeMessage)
-        })
-        .catch((error) =>
-          console.error(`Error fetching call details: ${error.message}`)
-        );
-    }
-
-    if (event === "media" && DataObject.media.track === "inbound") {
-      const rawAudio = Buffer.from(DataObject.media.payload, "base64");
-      this.deepgram.send(rawAudio);
-    }
-
-    if (event === "stop") {
-      this.handleClose();
-    }
-
-    if (event === "mark") {
-      console.log("twilio: Mark event received", data);
-    }
-  };
-
-  private handleClose() {
-    log("The call is closed");
-  }
-
-  getAgentObject() {
-    return this.AgentObject;
-  }
-
-  sendData(messageJSON: string) {
-    // Check if the connection is open before sending
-    if (this.connection.readyState === WebSocket.OPEN) {
-      try {
-        this.connection.send(messageJSON);
-        // console.log("Message sent successfully");
-      } catch (error) {
-        console.error("Error sending message:", error);
-      }
-    } else {
-      console.error("WebSocket connection is not open. Cannot send message.");
-    }
-  }
-}
-
-let llmStart = 0;
-let ttsStart = 0;
-let firstByte = true;
-let speaking = false;
-let send_first_sentence_input_time: number | null = null;
-const chars_to_check = [".", ",", "!", "?", ";", ":"];
-let keepAlive: any;
-
-const setupDeepgram = (mediaStream: any) => {
-  let is_finals: any[] = [];
-  const deepgram = deepgramClient.listen.live({
-    // Model
-    model: "nova-2-phonecall",
-    language: "en",
-    // Formatting
-    smart_format: true,
-    // Audio
-    encoding: "mulaw",
-    sample_rate: 8000,
-    channels: 1,
-    multichannel: false,
-    // End of Speech
-    no_delay: true,
-    interim_results: true,
-    endpointing: 300,
-    utterance_end_ms: 1000,
-  });
-
-  if (keepAlive) clearInterval(keepAlive);
-  keepAlive = setInterval(() => {
-    deepgram.keepAlive(); // Keeps the connection alive
-  }, 10 * 1000);
-
-  deepgram.addListener(LiveTranscriptionEvents.Open, async () => {
-    console.log("deepgram STT: Connected");
-
-    deepgram.addListener(LiveTranscriptionEvents.Transcript, (data) => {
-      const transcript = data.channel.alternatives[0].transcript;
-      if (transcript !== "") {
-        if (data.is_final) {
-          is_finals.push(transcript);
-          if (data.speech_final) {
-            const utterance = is_finals.join(" ");
-            is_finals = [];
-            log(`deepgram STT: [Speech Final] ${utterance}`);
-            llmStart = Date.now();
-            promptLLM(mediaStream, utterance); // Send the final transcript to OpenAI for response
-          } else {
-            log(`deepgram STT:  [Is Final] ${transcript}`);
-          }
-        } else {
-          log(`deepgram STT:    [Interim Result] ${transcript}`);
-          if (speaking) {
-            console.log("twilio: clear audio playback", streamSid);
-            // Handles Barge In
-            const messageJSON = JSON.stringify({
-              event: "clear",
-              streamSid: streamSid,
-            });
-            mediaStream.sendData(messageJSON);
-
-            mediaStream.deepgramTTSWebsocket.send(
-              JSON.stringify({ type: "Clear" })
-            );
-            speaking = false;
-          }
-        }
-      }
-    });
-
-    deepgram.addListener(LiveTranscriptionEvents.UtteranceEnd, (data) => {
-      if (is_finals.length > 0) {
-        console.log("deepgram STT: [Utterance End]");
-        const utterance = is_finals.join(" ");
-        is_finals = [];
-        console.log(`deepgram STT: [Speech Final] ${utterance}`);
-        llmStart = Date.now();
-        promptLLM(mediaStream, utterance)
-      }
-    });
-
-    deepgram.addListener(LiveTranscriptionEvents.Close, async () => {
-      log("deepgram STT: disconnected");
-      clearInterval(keepAlive);
-      deepgram.requestClose();
-    });
-
-    deepgram.addListener(LiveTranscriptionEvents.Error, async (error) => {
-      console.log("deepgram STT: error received");
-      console.error(error);
-    });
-
-    deepgram.addListener(LiveTranscriptionEvents.Metadata, (data) => {
-      // console.log('deepgram STT: metadata received:', data)
-    });
-  });
-
-  return deepgram;
-};
-
-async function promptLLM(mediaStream: any, prompt: string) {
-  const AgentObject = mediaStream.getAgentObject();
-  const { agentPrompt } = AgentObject[0];
-  console.log(`agentPrompt : ${agentPrompt}`);
-
-  const stream = openai.beta.chat.completions.stream({
-    model: "gpt-3.5-turbo",
-    stream: true,
-    messages: [
-      {
-        role: "assistant",
-        content: agentPrompt,
-      },
-      {
-        role: "user",
-        content: prompt,
-      },
-    ],
-  });
-
-  speaking = true;
-  let firstToken = true;
-  for await (const chunk of stream) {
-    if (speaking) {
-      if (firstToken) {
-        const end = Date.now();
-        const duration = end - llmStart;
-        ttsStart = Date.now();
-        console.warn(
-          "\n>>> openai LLM: Time to First Token = ",
-          duration,
-          "\n"
-        );
-        firstToken = false;
-        firstByte = true;
-      }
-      let chunk_message = chunk.choices[0].delta.content;
-      if (chunk_message) {
-        process.stdout.write(chunk_message);
-        console.log(`GPT response: ${chunk_message}`);
-        if (
-          !send_first_sentence_input_time &&
-          containsAnyChars(chunk_message)
-        ) {
-          send_first_sentence_input_time = Date.now();
-        }
-        mediaStream.deepgramTTSWebsocket.send(
-          JSON.stringify({ type: "Speak", text: chunk_message })
-        );
-      }
-    }
-  }
-  // Tell TTS Websocket we're finished generating tokens
-  mediaStream.deepgramTTSWebsocket.send(JSON.stringify({ type: "Flush" }));
-}
-function containsAnyChars(str: string): boolean {
-  // Convert the string to an array of characters
-  let strArray: string[] = Array.from(str);
-
-  // Check if any character in strArray exists in chars_to_check
-  return strArray.some((char: string) => chars_to_check.includes(char));
-}
-
-const deepgramTTSWebsocketURL =
-  "wss://api.deepgram.com/v1/speak?encoding=mulaw&sample_rate=8000&container=none&model=aura-orpheus-en";
-
-const setupDeepgramWebsocket = (mediaStream: any) => {
-  const options = {
-    headers: {
-      Authorization: `Token ${process.env.DEEPGRAM_API_KEY}`,
-    },
-  };
-  const ws = new WebSocketClient(deepgramTTSWebsocketURL, options);
-
-  ws.on("open", function open() {
-    console.log("deepgram TTS: Connected");
-  });
-
-  ws.on("message", function incoming(data) {
-    // Handles barge in
-    if (speaking) {
-      try {
-        let json = JSON.parse(data.toString());
-        console.log("deepgram TTS: ", data.toString());
-        return;
-      } catch (e) {
-        // Ignore
-      }
-      if (firstByte) {
-        const end = Date.now();
-        const duration = end - ttsStart;
-        console.warn(
-          "\n\n>>> deepgram TTS: Time to First Byte = ",
-          duration,
-          "\n"
-        );
-        firstByte = false;
-        if (send_first_sentence_input_time) {
-          console.log(
-            `>>> deepgram TTS: Time to First Byte from end of sentence token = `,
-            end - send_first_sentence_input_time
-          );
-        }
-      }
-      const payload = data.toString("base64");
-
-      const messageJSON = JSON.stringify({
-        event: "media",
-        streamSid,
-        media: { payload },
-      });
-      mediaStream.sendData(messageJSON);
-    }
-  });
-
-  ws.on("close", function close() {
-    console.log("deepgram TTS: Disconnected from the WebSocket server");
-  });
-
-  ws.on("error", function error(error) {
-    console.log("deepgram TTS: error received");
-    console.error(error);
-  });
-  return ws;
-};
-
-// Function to broadcast logs to WebSocket clients
-function log(message: string) {
-  wss.clients.forEach(
-    (client: { readyState: any; OPEN: any; send: (arg0: string) => void }) => {
-      if (client.readyState === client.OPEN) {
-        client.send(message);
-      }
-    }
-  );
-}
-
-// Set a timeout to call the function from the separate file
 
 // Welcome route
 app.get("/", (req: Request, res: Response) => {
@@ -463,7 +108,7 @@ app.post("/status-callback", (req: Request, res: Response) => {
   console.log(eventMessage);
 
   // Broadcast the event to WebSocket clients
-  log(eventMessage);
+  console.log(eventMessage);
 
   res.status(200).send("Status callback received");
 });
